Clean Up of Code:
- Made it so HarEntry, HarEntry.Request, and HarEntry.Response act like dicts
- Changed type evaluations to isinstance
- Rename sub_classes to http
- Added Mixins for shared functions
- Removed try/expect where appropriate
